made the custom filter tag modular
